requirements.txt
numpy
PyOpenGL
PyOpenGL-accelerate
pil

